feat: Lab2 notebook for Zigbee sniffing
Updates the Lab2 notebook to focus on sniffing distributed Zigbee networks.

This includes changes to the lab objectives, challenge, and walk-through,
revising the content to reflect Zigbee-specific concepts and procedures,
such as identifying cluster and endpoint addresses.

It also adds instructions for flashing the CC1352 with the multiprotocol sniffer firmware.

Adds images to the markdown to illustrate the wireshark config, as well
as the sniffer in execution.